requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
langchain-core